Created ManifestWriter interface Added ManifestWriter to BaseModule Use CucumberManifestWriter in cucumber tests
